requirements.txt
streamlit
yfinance
pandas
python-pptx
matplotlib
feedparser